cria helper de requisição da api
